+++ strucmotif-search-update/src/main/java/org/rcsb/strucmotif/update/UpdateItem.java
@@ -1,22 +1,33 @@
 package org.rcsb.strucmotif.update;
 
+import org.rcsb.strucmotif.io.StructureDataProvider;
+
 import java.net.URL;
 
 /**
  * One piece of the update. Usually just references a PDB entry.<p>
- * Can also point to external resources or local files via a URL. In that case, you must make sure to provide a unique,
- * yet compact structureIdentifier.
+ * Can also point to external resources or local files via a URL. In that case, you must make sure to provide a unique
+ * structureIdentifier.
  */
 public class UpdateItem {
     private final String structureIdentifier;
     private final URL url;
+    private final int modelIdentifier;
 
     /**
      * Update by PDB-ID.
      * @param structureIdentifier a PDB entry ID
      */
     public UpdateItem(String structureIdentifier) {
-        this(structureIdentifier, null);
+        this(structureIdentifier, null, StructureDataProvider.DEFAULT_MODEL_IDENTIFIER);
+    }
+
+    /**
+     * Update a specific model of a PDB-ID.
+     * @param structureIdentifier a PDB entry ID
+     */
+    public UpdateItem(String structureIdentifier, int modelIdentifier) {
+        this(structureIdentifier, null, modelIdentifier);
     }
 
     /**
@@ -25,8 +36,19 @@ public class UpdateItem {
      * @param url data source pointing to non-PDB structures, either by URL to some external resource or to a local file
      */
     public UpdateItem(String structureIdentifier, URL url) {
+        this(structureIdentifier, url, StructureDataProvider.DEFAULT_MODEL_IDENTIFIER);
+    }
+
+    /**
+     * Update by identifier and external URL, while being model-aware.
+     * @param structureIdentifier the unique ID this item will have
+     * @param url data source pointing to non-PDB structures, either by URL to some external resource or to a local file
+     * @param modelIdentifier int value of the model to index
+     */
+    public UpdateItem(String structureIdentifier, URL url, int modelIdentifier) {
         this.structureIdentifier = structureIdentifier;
         this.url = url;
+        this.modelIdentifier = modelIdentifier;
     }
 
     /**
@@ -43,5 +65,13 @@ public class UpdateItem {
      */
     public URL getUrl() {
         return url;
+    }
+
+    /**
+     * Which model to index?
+     * @return an int (usually, 1)
+     */
+    public int getModelIdentifier() {
+        return modelIdentifier;
     }
 }